requirements.txt
lmdb==0.97
numpy==1.17.2
Pillow==6.1.0
six==1.12.0
torch==1.7.1
torchvision==0.8.2
